Add files via upload
This is the v2 of the simple Ceasar Cipher with Python. This version encrypts only the 26 English alphabets. It can also retain information about the positions of space and upper-case characters. You can type a proper message in English (without any special symbols). The program with encrypt and decrypt your message provided that you give it the correct encryption key.